requirements.txt
alembic==1.14.1
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
django-celery-beat>=2.6.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.8.0
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-extra-fields==3.7.0
drf-yasg==1.21.9
filetype==1.2.0
fonttools==4.56.0
inflection==0.5.1
kombu==5.4.2
Mako==1.3.9
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.0
webencodings==0.5.1
zopfli==0.2.3.post1
celery==5.4.0
openpyxl==3.1.5
